minor reorg, added asynchronous zipper (just parallel with 2 zippers)
